Change dtype of tensor

+++ src/reader.py
@@ -10,7 +10,7 @@ def epoch_size(raw_data, batch_size, num_steps):
 def img_producer(raw_data, batch_size, num_steps, shuffle=False, name="IMGProducer"):
     with tf.name_scope(name):
         N, rows, cols = raw_data.shape
-        raw_data = tf.convert_to_tensor(raw_data, name="raw_data", dtype=tf.int32)
+        raw_data = tf.convert_to_tensor(raw_data, name="raw_data", dtype=tf.float32)
 
         samples_per_image = (cols - 1) / num_steps
         epoch_size = N*samples_per_image / batch_size
